Various performance improvements

+++ FrameCounter.cs
@@ -2,16 +2,16 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 public class FrameCounter{
-    public float AverageFramesPerSecond { get; private set; }
-    private readonly Queue<float> _sampleBuffer = new();
+    public double AverageFramesPerSecond { get; private set; }
+    private readonly Queue<double> _sampleBuffer = new();
     public FrameCounter(){
         for(int i=0;i<10;i++){
             _sampleBuffer.Enqueue(0);
         }
     }
-    public void Update(float deltaTime){
+    public void Update(double deltaTime){
         _sampleBuffer.Dequeue();
         _sampleBuffer.Enqueue(deltaTime);
-        AverageFramesPerSecond = (float)Math.Round(1.0f/_sampleBuffer.Average(),2);
+        AverageFramesPerSecond = Math.Round(1.0d/_sampleBuffer.Average(),2);
     }
 }
